% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_amr.R
\name{amr_mgen}
\alias{amr_mgen}
\title{Development of AMR}
\usage{
amr_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
either due to treatment or spontaneously(?)
}
