% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clinical.R
\name{clinical_helpers}
\alias{clinical_helpers}
\alias{set_event_time_rnorm}
\alias{flag_ids_for_recovery}
\alias{flag_ids_for_treatment_failure}
\alias{update_amr_status}
\alias{update_attrs_for_new_treatment}
\title{Clinical Helper Functions for Treatment and AMR Management}
\usage{
set_event_time_rnorm(dat, at, ids, event_name, param1, param2 = NULL)

flag_ids_for_recovery(
  dat,
  at,
  ids,
  tx_success_attr = "tx_success",
  tx_success_val = 1
)

flag_ids_for_treatment_failure(
  dat,
  at,
  ids,
  tx_success_attr = "tx_success",
  tx_success_val = 0
)

update_amr_status(dat, at, ids, amr_attr, amr_present_val = 1)

update_attrs_for_new_treatment(
  dat,
  at,
  ids,
  treatment_type,
  curr_tx_attr = "curr_tx",
  tx_end_day_attr = "tx_end_day",
  tx_success_attr = "tx_success",
  tx_return_day_attr = "tx_return_day"
)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{EpiModel::netsim}}.}

\item{at}{Current time step.}

\item{ids}{A vector of individual IDs for whom the event or treatment is being set or updated.}

\item{event_name}{The name of the event attribute to be updated for the specified individuals.}

\item{param1}{The mean of the normal distribution from which the event times are drawn.}

\item{param2}{The standard deviation of the normal distribution from which the event times are drawn. Defaults to 1.}

\item{tx_success_attr}{The name of the attribute indicating treatment success. Defaults to \code{"tx_success"}.}

\item{tx_success_val}{The value to set for treatment failure. Defaults to \code{0}.}

\item{amr_attr}{The name of the attribute indicating AMR status.}

\item{amr_present_val}{The value to set for individuals with AMR. Defaults to \code{1}.}

\item{treatment_type}{The type of treatment being administered. Should be one of \code{"doxy"}, \code{"moxi"}, \code{"az"}, or \code{"other"}.}

\item{curr_tx_attr}{The name of the attribute indicating the current treatment. Defaults to \code{"curr_tx"}.}

\item{tx_end_day_attr}{The name of the attribute indicating the treatment end day. Defaults to \code{"tx_end_day"}.}

\item{tx_return_day_attr}{The name of the attribute indicating the return day for treatment. Defaults to \code{"tx_return_day"}.}
}
\value{
The updated \code{dat} object with modified attributes based on the specified event or treatment.
}
\description{
Sets the time for a specified event for a given set of individuals,
drawing from a normal distribution with the specified mean and standard deviation.

Flags the specified individuals as having recovered from treatment.

Flags the specified individuals as having experienced treatment failure.

Updates the antimicrobial resistance (AMR) status for the specified individuals.

Updates the attributes for individuals receiving a new treatment, including current treatment, treatment end day, treatment success, and return day.
}
