% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network_targets.R
\name{generate_init_network}
\alias{generate_init_network}
\title{Generate Starting Network for ERGM Estimation}
\usage{
generate_init_network(
  params,
  seed = NULL,
  assign_deg_casual = FALSE,
  assign_deg_main = FALSE
)
}
\arguments{
\item{params}{a list of parameters with a sublist called "pop"
with the size of the network
and the named attributes and their distribution in the population}

\item{seed}{optional, maintains distribution of nodal attributes}

\item{assign_deg_casual}{default = FALSE, if TRUE, uses additional
parameters information in "params" to set
a plausible casual degree for each node
(can help fit main partnership network when cross-network terms present)}

\item{assign_deg_main}{default = FALSE, if TRUE, uses additional
parameters information in "params" to set
a plausible main degree for each node
(can help fit casual partnership network when cross-network terms present)
when casual network fit first in workflow)}
}
\value{
an object of class network
}
\description{
Generates empty network of specified size
and adds nodal attributes.
Currently these attributes are specific to this project:
a sex variable ("female"), race, age group, age
}
