% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_simulated_rels.R
\name{get_mean_durations}
\alias{get_mean_durations}
\title{Get Mean Durations of Relationships at End of Simulation}
\usage{
get_mean_durations(sim, nets = c("main", "casual"), yaml_params_loc)
}
\arguments{
\item{sim}{A simulation object of class \code{EpiModel::netsim}.}

\item{nets}{A character vector specifying the networks to calculate durations for, default is c("main", "casual").}

\item{yaml_params_loc}{Path to the YAML file containing target durations.}
}
\value{
A data frame summarizing the target and simulated mean durations for each network,
along with the standard deviation of the simulated durations.
}
\description{
Calculates the mean durations of relationships in the main and casual networks at the end
of the simulation, comparing them to target durations specified in a YAML file.
}
