% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimodel_utilities.R
\name{get_successful_ids_binom}
\alias{get_successful_ids_binom}
\title{Helper Function for Binomial Sampling of Eligible IDs}
\usage{
get_successful_ids_binom(elig_ids, eval_prob)
}
\arguments{
\item{elig_ids}{A vector of eligible IDs (e.g., indices of symptomatic infected individuals).}

\item{eval_prob}{A vector of probabilities corresponding to each eligible ID, or a single
probability to be applied to all eligible IDs.}
}
\value{
A vector of IDs or an empty integer vector.
}
\description{
This function takes a vector of eligible IDs and a corresponding
vector of probabilities, and returns a vector of IDs that are "successful"
based on a binomial draw. If there are no eligible IDs, it returns an empty integer vector.
}
