% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_simulated_rels.R
\name{get_target_degrees}
\alias{get_target_degrees}
\title{Get Target Mean Degrees for Main and Casual Networks}
\usage{
get_target_degrees(yaml_params_loc, nets = c("main", "casual"))
}
\arguments{
\item{yaml_params_loc}{Path to the YAML file containing network targets.}

\item{nets}{Character vector specifying which networks to extract targets
for.
Default is c("main", "casual").}
}
\value{
A data frame where each cell represents the mean degree that
set of attrs and network type.
}
\description{
Extracts target degrees for main and casual networks
from a YAML file. Currently assumes the YAML file
includes sexual activity terms under "nodefactor" for each network
stratified by a joint set of attributes,
kept as vectors under "prediction_pop"
}
