% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_simulated_rels.R
\name{get_target_degrees_age_race}
\alias{get_target_degrees_age_race}
\title{Get Target Mean Degrees for Main and Casual Networks by Age and Race}
\usage{
get_target_degrees_age_race(
  yaml_params_loc,
  nets = c("main", "casual"),
  joint_attrs = c("age", "race")
)
}
\arguments{
\item{yaml_params_loc}{Path to the YAML file containing network targets.}

\item{nets}{Character vector specifying which networks to extract targets for. Default is c("main", "casual").}

\item{joint_attrs}{Character vector specifying the two joint attribute to extract targets for.
Default is c("age", "race").}
}
\value{
A data frame where each cell represents the mean degree for that age, race, and network type.
}
\description{
Extracts target degrees for main and casual networks from a YAML file. Currently assumes the YAML file
includes sexual activity terms under "nodefactor" for each network stratified by a joint set of attributes,
age and race, and description of attributes distributions under a sublist named "pop".
}
