% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_e_to_i.R
\name{mod_ei_mgen}
\alias{mod_ei_mgen}
\title{Transition from Exposed to Infected}
\usage{
mod_ei_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{EpiModel::netsim}}.}

\item{at}{Current time step.}
}
\description{
Evaluates eligible exposed persons for infection
based on sex-specific rates, updates infection status accordingly
}
