% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infection.R
\name{mod_infection}
\alias{mod_infection}
\title{Model Infection between Partners}
\usage{
mod_infection(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
Disease transmission module accounting for age-specific activity rates
}
