% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_initialize.R
\name{mod_initialize_mgen}
\alias{mod_initialize_mgen}
\alias{init_mgen_status}
\title{Initialization Module}
\usage{
mod_initialize_mgen(x, param, init, control, s)

init_mgen_status(dat)
}
\arguments{
\item{x}{If \code{control$start == 1}, either a fitted network model object
of class \code{netest} or a list of such objects. If
\code{control$start > 1}, an object of class \code{netsim}. When
multiple networks are used, the node sets (including network size
and nodal attributes) are assumed to be the same for all networks.}

\item{param}{An \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\value{
A \code{netsim_dat} class main data object.

A modified \code{dat} object with initialized infections
}
\description{
Initialize networks, starting nodal attributes,
infection status,and other necessary components for STI
transmission models. Most of this module is adapted from
\code{EpiModel}'s \code{initialize.net} function, with bespoke
checks for required attributes and infection assignment (in future)

Initialize infection status and related attributes for STI
transmission models. This function is called within \code{mod_initialize_mgen}
to set up the initial infection status of the population based on the number
of initial infections specified in the \code{init} object.
}
\details{
When re-initializing a simulation, the \code{netsim} object passed
to \code{initialize.net} must contain the elements \code{param},
\code{nwparam}, \code{epi}, \code{coef.form}, and \code{num.nw}.
}
