% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recovery.R
\name{mod_recovery_mgen}
\alias{mod_recovery_mgen}
\title{Recover from M. genitalium infection}
\usage{
mod_recovery_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{EpiModel::netsim}}.}

\item{at}{Current time step.}
}
\description{
Evaluates eligible infected persons for recovery
based on sex-specific rates, resets infection attrs if recovered
}
