% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sti_initialize.R
\name{mod_sti_initialize}
\alias{mod_sti_initialize}
\title{Initialization Module}
\usage{
mod_sti_initialize(x, param, init, control, s)
}
\arguments{
\item{x}{If \code{control$start == 1}, either a fitted network model object
of class \code{netest} or a list of such objects. If
\code{control$start > 1}, an object of class \code{netsim}. When
multiple networks are used, the node sets (including network size
and nodal attributes) are assumed to be the same for all networks.}

\item{param}{An \code{EpiModel} object of class \code{\link{param.net}}.}

\item{init}{An \code{EpiModel} object of class \code{\link{init.net}}.}

\item{control}{An \code{EpiModel} object of class \code{\link{control.net}}.}

\item{s}{Simulation number, used for restarting dependent simulations.}
}
\value{
A \code{netsim_dat} class main data object.
}
\description{
Initialize networks, starting nodal attributes, infection status,
and other necessary components for STI transmission models. Most of this module
is adapted from \code{EpiModel}'s \code{initialize.net} function, with bespoke
checks for required attributes and infection assigment (in future)
}
\details{
When re-initializing a simulation, the \code{netsim} object passed
to \code{initialize.net} must contain the elements \code{param},
\code{nwparam}, \code{epi}, \code{coef.form}, and \code{num.nw}.
}
\keyword{internal}
