% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_simulated_rels.R
\name{plot_edges_history}
\alias{plot_edges_history}
\title{Plot Edges History}
\usage{
plot_edges_history(x, network, type)
}
\arguments{
\item{x}{Either a data frame containing the edges history (the output of \code{get_edges_history()}),
or a simulation object of class \code{EpiModel::netsim}.
If a simulation object is provided, edges history will be extracted using \code{get_edges_history()}.}

\item{network}{A character string specifying the network type, either "main" or "casual".}

\item{type}{A character string specifying the type of difference to plot, either "percent", "absolute", or "edges".}
}
\value{
A ggplot object showing the edges history over time for the specified network and type.
}
\description{
Plots the edges history for a specified network and type of difference (absolute, percent, or edges).
}
