% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_progression.R
\name{progression_mgen}
\alias{progression_mgen}
\title{Mgen Disease Progression & Recovery}
\usage{
progression_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
Mgen disease progression
}
