% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infection.R
\name{set_infection_attrs_mgen}
\alias{set_infection_attrs_mgen}
\title{Assigns Infection-Related Attributes for Newly Exposed Individuals (M. genitalium)}
\usage{
set_infection_attrs_mgen(dat, at, ids_new_e)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{EpiModel::netsim}}.}

\item{at}{Current time step.}

\item{ids_new_e}{A vector of IDs corresponding to newly exposed individuals.}
}
\description{
Assigns symptomatic status, calculates the time of transition from exposed
to infected based on the incubation period, calculates time of natural recovery.
This module should be run after the infection module in the same time step.
Note: This module assumes a normal distribution for the incubation period and infection duration,
with the mean specified in the parameters and a standard deviation equal to half of the mean.
}
