% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_simulated_rels.R
\name{summarize_final_degrees}
\alias{summarize_final_degrees}
\title{Summarize Final Degrees from Simulation}
\usage{
summarize_final_degrees(input, network)
}
\arguments{
\item{input}{An object of class \code{EpiModel::netest}, \code{EpiModel::netsim} or \code{EpiModel::netdx}.}
}
\value{
A data frame summarizing the mean degree, interquartile range (IQR), and data source
for each age and race combination
}
\description{
Summarizes the final degrees of individuals in the main and casual networks
at the end of the simulation and calculates the mean degree for each age and race combination.
For \code{EpiModel::netest} objects, a network is simulated using the "newnetwork" slot as its basis.
For \code{EpiModel::netdx} objects, data is extracted from the "tedgelist" (timed edgelist) slot.
For \code{EpiModel::netsim} objects, data is extracted from the "network" slot.
}
