% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network_targets.R
\name{targets}
\alias{targets}
\alias{calc_nodefactor}
\alias{calc_nodematch}
\alias{calc_edges}
\alias{calc_absdiff}
\alias{calc_concurrent}
\alias{calc_cross_network}
\alias{calc_nodecov_age}
\alias{check_targets}
\alias{check_conditions}
\title{Target Stats Calculation Helpers}
\usage{
calc_nodefactor(params, nw_attrs, rel, grouping_vars = NULL)

calc_nodematch(params, attr_name, nf_targets, rel, diff)

calc_edges(params, nw_attrs, rel, grouping_vars = NULL)

calc_absdiff(params, rel, count_type, edges)

calc_concurrent(params, rel, num)

calc_cross_network(params, rel, edges)

calc_nodecov_age(params, rel, attr_name, edges, attr_squared = FALSE)

check_targets(edges, final_targets, count_type, threshold = 0.01)

check_conditions(nw, params, rel, count_type, attr_name)
}
\arguments{
\item{params}{the parameter list object with parameters information for
all networks estimated from data}

\item{nw_attrs}{list of network attribute vectors}

\item{rel}{string, which relationship/network targets are calculated}

\item{grouping_vars}{attributes to group results by, NULL for all edges}

\item{attr_name}{string, used for nodefactor and nodematch targets.}

\item{nf_targets}{calculated from calc_single_attr_nodefactor()}

\item{diff}{default = FALSE, if TRUE, calc nodematch targets among each group}

\item{count_type}{string, which ERGM term to generate targets
Currently can only be in the form:
"edges", "nodefactor", "nodematch", "absdiff_sqrt_age", or "concurrent".}

\item{edges}{output from calc_edges()}

\item{num}{the number of nodes in the network (population size)}

\item{attr_squared}{for nodecov target calculation,
should squared version of attribute be calculated? (usually, age)}

\item{final_targets}{vector, final ergm term targets
to be checked before output}

\item{threshold}{default = 0.01, proportion of expected
activity based on edges that
calulated target is allowed within (+/- threshold)}

\item{nw}{the network object outputted from "generate_init_network"
(usually already in environment during workflow)}
}
\description{
Small helper functions used as part of calc_targets()
}
