% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_testing.R
\name{testing_mgen}
\alias{testing_mgen}
\title{Mgen Testing and Screening}
\usage{
testing_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
Regular screening and symptomatic testing
}
