% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_edge_tracker.R
\name{track_edges}
\alias{track_edges}
\title{Track Number of Relationships in Each Network}
\usage{
track_edges(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{EpiModel::netsim}}.}

\item{at}{Current time step.}
}
\description{
This module tracks the number of edges in each network at each
time step, storing the counts in the epidemiological summary statistics list.
}
