% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgen_tx.R
\name{tx_mgen}
\alias{tx_mgen}
\title{Mgen Treatment}
\usage{
tx_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
treatment of mgen
}
