% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_simulated_rels.R
\name{get_edges_history}
\alias{get_edges_history}
\title{Get Edges History from Simulation}
\usage{
get_edges_history(
  sim,
  edge_prefix = "edges_net_",
  net_names = c("main", "casual", "inst")
)
}
\arguments{
\item{sim}{A simulation object of class \code{EpiModel::netsim}.}

\item{edge_prefix}{A character string specifying the prefix used for edge
history in the simulation object. Default is "edges_net_".}

\item{net_names}{A character vector specifying the names of the networks
corresponding to the edge history in the simulation object. Default is
c("main", "casual", "inst").}
}
\value{
A df with time, simulation id, edges for all networks,
the difference from target edges, and the percentage diff from target edges.
}
\description{
Extracts edges history as the number of edges per time step from
a simulation object, calculating the difference between the simulated edges
and the target edges for all networks.
Assumes edge history is tracked in the simulation object via the "epi" list,
not via setting the control parameter 'nwstats' TRUE.
}
