% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_simulated_rels.R
\name{get_edges_history}
\alias{get_edges_history}
\title{Get Edges History from Simulation}
\usage{
get_edges_history(sim, nets = c("main", "casual"))
}
\arguments{
\item{sim}{A simulation object of class \code{EpiModel::netsim}.}

\item{nets}{Character vector specifying which networks to extract targets for. Default is c("main", "casual").}
}
\value{
A data frame with time, simulation identifier, edges for main and casual networks,
the difference from target edges, and the percentage difference from target edges.
}
\description{
Extracts edges history as the number of edges per time step from a simulation object,
calculating the difference between the simulated edges and the target edges for both main and casual networks.
NOTE: Assumes edge history is tracked in the simulation object in the "epi" list, not via setting the control
parameter 'nwstats' TRUE.
}
