% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clinical.R
\name{mod_clinical_mgen}
\alias{mod_clinical_mgen}
\title{Module for clinical aspects of M. genitalium infection}
\usage{
mod_clinical_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This module handles the clinical aspects of M. genitalium infection,
including screening, testing, diagnosis, and treatment outcomes
depending on the specified scenario.
}
