% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vital_dynamics.R
\name{vitals}
\alias{vitals}
\alias{mod_aging_mgen}
\alias{mod_departures_mgen}
\alias{mod_arrivals_mgen}
\title{Modules for vital dynamics}
\usage{
mod_aging_mgen(dat, at)

mod_departures_mgen(dat, at)

mod_arrivals_mgen(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{EpiModel::netsim}}.}

\item{at}{Current time step.}
}
\description{
Handles node aging, departure, and arrivals
}
