% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vital_dynamics.R
\name{vitals}
\alias{vitals}
\alias{mod_aging}
\alias{mod_departures}
\alias{mod_arrivals}
\title{Modules for vital dynamics}
\usage{
mod_aging(dat, at)

mod_departures(dat, at)

mod_arrivals(dat, at)
}
\arguments{
\item{dat}{Main \code{netsim_dat} object containing a \code{networkDynamic}
object and other initialization information passed from
\code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
Handles node aging, departure, and arrivals
}
